rename common.js to common-evals.js
