+++ server/src/trade/buy/strategies/common/common-evals.js
